@@ -1,20 +1,15 @@
-import { State, Action, StateContext, Selector, NgxsOnInit } from '@ngxs/store';
-import { InkAppSettingsItem, InkAppSettings } from '@lib/models';
+import { InkAppSettings } from '@lib/models';
+import { InkSettingsService } from '@lib/services';
+import { Action, NgxsOnInit, Selector, State, StateContext } from '@ngxs/store';
+import { DEFAULT_SETTINGS } from '@root/ink.config';
 import {
-  UpdateSettingsItem,
-  PopulateSettings,
-  MergeSettings,
-  FetchRemoteGist,
-  CreateRemoteGist,
-  UpdateRemoteGist,
   AddSettingsItem,
   CreateSettingsItem,
-  PopulateDefaultSettings
-} from '@store/actions';
-import { SelectMultipleControlValueAccessor } from '@angular/forms';
-import { InkSettingsService, InkGistService } from '@lib/services';
-import { tap, map } from 'rxjs/operators';
-import { DEFAULT_SETTINGS } from '@root/ink.config';
+  MergeSettings,
+  PopulateDefaultSettings,
+  PopulateSettings,
+  UpdateSettingsItem } from '@store/actions';
+import { map } from 'rxjs/operators';
 
 @State<Partial<InkAppSettings>>({
   name: 'settings',
@@ -26,10 +21,10 @@ export class SettingsState implements NgxsOnInit {
     return state.filter(s => s.key === 'view')[0].value;
   }
 
-  constructor(private _service: InkSettingsService, private _gistService: InkGistService) {}
+  constructor(private _service: InkSettingsService) {}
 
   async ngxsOnInit(ctx: StateContext<InkAppSettings>) {
-    const defaultSettings = await this._service.getAll();
+    const defaultSettings = this._service.getAll();
     if (defaultSettings.length === 0) {
       return ctx.dispatch(new PopulateDefaultSettings(DEFAULT_SETTINGS));
     }
@@ -38,7 +33,7 @@ export class SettingsState implements NgxsOnInit {
 
   @Action(UpdateSettingsItem)
   updateSettings(ctx: StateContext<InkAppSettings>, action: UpdateSettingsItem) {
-    return this._service.update(action.key, action.value).then((doc: any) => {
+    return this._service.update(action.key, action.value).pipe(map((doc: any) => {
       const state = ctx.getState();
       const newState = state.map(s => {
         if (s.key === action.key) {
@@ -47,14 +42,14 @@ export class SettingsState implements NgxsOnInit {
         return s;
       });
       ctx.setState([...newState]);
-    });
+    }));
   }
 
   @Action(CreateSettingsItem)
   createSettingsItem(ctx: StateContext<InkAppSettings>, action: CreateSettingsItem) {
-    this._service.add(action.key, action.value).then(settingsItem => {
+    this._service.add(action.key, action.value).pipe(map(settingsItem => {
       ctx.dispatch(new AddSettingsItem({ key: action.key, value: action.value }));
-    });
+    }));
   }
 
   @Action(AddSettingsItem)
@@ -65,9 +60,9 @@ export class SettingsState implements NgxsOnInit {
 
   @Action(PopulateDefaultSettings)
   populateDefaultSettings(ctx: StateContext<InkAppSettings>, action: PopulateDefaultSettings) {
-    this._service.addAll(action.settings).then(settings => {
+    this._service.addAll(action.settings).pipe(map(settings => {
       ctx.dispatch(new PopulateSettings(settings));
-    });
+    }));
   }
 
   @Action(PopulateSettings)
@@ -78,24 +73,5 @@ export class SettingsState implements NgxsOnInit {
   mergeSettings(ctx: StateContext<InkAppSettings>, action: MergeSettings) {
     const state = ctx.getState();
     ctx.setState([...state, ...action.settings]);
-  }
-
-  @Action(FetchRemoteGist)
-  fetchRemoteGist(ctx: StateContext<InkAppSettings>, action: FetchRemoteGist) {
-    return this._gistService.get();
-  }
-
-  @Action(CreateRemoteGist)
-  createRemoteGist(ctx: StateContext<InkAppSettings>, action: CreateRemoteGist) {
-    return this._gistService
-      .create(action.gistData)
-      .pipe(tap(gist => ctx.dispatch(new CreateSettingsItem('gist', gist))));
-  }
-
-  @Action(UpdateRemoteGist)
-  updateRemoteGist(ctx: StateContext<InkAppSettings>, action: UpdateRemoteGist) {
-    return this._gistService
-      .edit(action.gistId, action.gistData)
-      .pipe(tap(gist => ctx.dispatch(new UpdateSettingsItem('gist', gist))));
   }
 }
