@@ -1,4 +1,4 @@
-import { Component, OnInit, NgZone } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'inkapp-root',
@@ -7,7 +7,4 @@ import { Component, OnInit, NgZone } from '@angular/core';
   <router-outlet></router-outlet>
   `
 })
-export class InkApp implements OnInit {
-  constructor() {}
-  ngOnInit() {}
-}
+export class InkApp {}
