@@ -1,29 +1,12 @@
-export const DB_CONFIG = {
-  name: 'inkdb',
-  adapter: 'idb',
-  password: 'supercoolpassword',
-  multiInstance: true
-};
-
-export const DB_REMOTE_SYNC_URL = `http://${window.location.hostname}:10101/`;
-
 export const BOARD_DEFAULT_NAME = 'Beautiful';
 export const BOARD_DEFAULT_DESC = 'Default Board';
 
-export const DEFAULT_BUCKET_NAME = `Bucket #`;
-
-export const GITHUB_ACCESS_TOKEN_NAME = 'inkapp_access_token';
-export const GIST_NAME = 'INKAPP :: SYNC SETTINGS';
+export const BUCKET_DEFAULT_NAME = `Bucket #`;
 
 export const DEFAULT_COLOR = '#ffffff';
-
 export const DEFAULT_SETTINGS = [
   {
     key: 'view',
     value: 'default'
-  },
-  {
-    key: 'gist',
-    value: null
   }
 ];
