@@ -1,13 +1,11 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { Store, Select } from '@ngxs/store';
-import { Observable } from 'rxjs';
-import { map, filter, tap } from 'rxjs/operators';
-import { InkBoardMeta, InkBucketMeta, InkBuckets, InkAppSettings, InkAppSettingsItem } from '@lib/models';
-import { InkBucketsService } from '@lib/services';
-import { BucketsState, SettingsState } from '@store/states';
+import { Component, Input, OnInit } from '@angular/core';
+import { InkBoardMeta, InkBuckets } from '@lib/models';
 import { children } from '@lib/operators';
+import { Select, Store } from '@ngxs/store';
+import { BUCKET_DEFAULT_NAME } from '@root/ink.config';
 import { CreateBucket } from '@store/actions';
-import { DEFAULT_BUCKET_NAME } from '@root/ink.config';
+import { SettingsState } from '@store/states';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'inkapp-board',
@@ -20,10 +18,10 @@ export class BoardComponent implements OnInit {
   constructor(private _store: Store) {}
 
   ngOnInit() {
-    this.buckets$ = this._store.select(s => s.buckets).pipe(children(this.data._id, 'boardId'));
+    this.buckets$ = this._store.select(s => s.buckets).pipe(children(this.data.id, 'boardId'));
   }
 
   createBucket() {
-    this._store.dispatch(new CreateBucket({ name: DEFAULT_BUCKET_NAME, boardId: this.data._id }));
+    this._store.dispatch(new CreateBucket({ name: BUCKET_DEFAULT_NAME, boardId: this.data.id}));
   }
 }
