@@ -1,12 +1,10 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { Store, Select } from '@ngxs/store';
-import { Observable } from 'rxjs';
-import { filter, tap, map } from 'rxjs/operators';
-import { InkBucketMeta, InkAppView, InkDropMeta, InkAppSettingsItem, InkDrops } from '@lib/models';
-import { InkBucketsService, InkDropsService } from '@lib/services';
-import { CreateDrop, UpdateBucket, DeleteBucket } from '@store/actions';
-import { SettingsState } from '@store/states';
+import { Component, Input, OnInit } from '@angular/core';
+import { InkBucketMeta, InkDropMeta, InkDrops } from '@lib/models';
 import { children } from '@lib/operators';
+import { InkBucketsService, InkDropsService } from '@lib/services';
+import { Store } from '@ngxs/store';
+import { CreateDrop, DeleteBucket, UpdateBucket } from '@store/actions';
+import { Observable } from 'rxjs';
 
 @Component({
   selector: 'inkapp-bucket',
@@ -23,11 +21,13 @@ export class BucketComponent implements OnInit {
   ) {}
 
   ngOnInit() {
-    this.drops$ = this._store.select(s => s.drops).pipe(children(this.data._id, 'bucketId'));
+    this.drops$ = this._store.select(s => s.drops).pipe(children(this.data.id, 'bucketId'));
   }
   newDrop() {
     const data: InkDropMeta = {
-      bucketId: this.data._id,
+      id: null,
+      createdBy: null,
+      bucketId: this.data.id,
       displayValue: '#ffffff',
       meta: {},
       name: '#ffffff'
@@ -35,10 +35,10 @@ export class BucketComponent implements OnInit {
     this._store.dispatch(new CreateDrop(data));
   }
   onTitleChange(name) {
-    this._store.dispatch(new UpdateBucket({ ...(this.data as any)._data, name }));
+    this._store.dispatch(new UpdateBucket({ ...this.data, name }));
   }
 
   deleteBucket() {
-    this._store.dispatch(new DeleteBucket(this.data._id));
+    this._store.dispatch(new DeleteBucket(this.data.id));
   }
 }
