@@ -1,10 +1,10 @@
 import { Component } from '@angular/core';
-import { Store } from '@ngxs/store';
-import { map, tap, filter } from 'rxjs/operators';
 import { InkBucketsService } from '@lib/services';
-import { CreateBucket } from '@store/actions';
+import { Store } from '@ngxs/store';
+import { BUCKET_DEFAULT_NAME } from '@root/ink.config';
+import { AuthService } from '@services/auth.service';
+import { CreateBucket, ResetUser } from '@store/actions';
 import { BoardsState } from '@store/states';
-import { DEFAULT_BUCKET_NAME } from '@root/ink.config';
 
 @Component({
   selector: 'inkapp-header',
@@ -12,9 +12,17 @@ import { DEFAULT_BUCKET_NAME } from '@root/ink.config';
 })
 export class HeaderComponent {
   currentBoardId: string;
-  constructor(private _store: Store, private _bucketService: InkBucketsService) {}
+  constructor(public auth: AuthService, private _store: Store, private _bucketService: InkBucketsService) {}
   newBucket() {
     const defaultBoardId = this._store.selectSnapshot(BoardsState.defaultBoard);
-    this._store.dispatch(new CreateBucket({ name: DEFAULT_BUCKET_NAME, boardId: defaultBoardId }));
+    this._store.dispatch(new CreateBucket({ name: BUCKET_DEFAULT_NAME, boardId: defaultBoardId }));
+  }
+  googleLogin() {
+    this.auth.googleLogin().subscribe(user => {
+      console.log('gogole login', user);
+    });
+  }
+  logout() {
+    this._store.dispatch(new ResetUser());
   }
 }
