@@ -1,4 +1,4 @@
-export { EditableTitleElement } from './editable-title/editable-title.element';
 export { ActionItemElement } from './action-item/action-item.element';
-export { UserCardElement } from './user-card/user-card.element';
 export { BillboardElement } from './billboard/billboard.element';
+export { EditableTitleElement } from './editable-title/editable-title.element';
+
