@@ -1,20 +1,8 @@
-import { State, Action, StateContext } from '@ngxs/store';
 import { InkDrops } from '@lib/models';
-import {
-  FetchAllDrops,
-  PopulateAllDrops,
-  UpdateDrop,
-  PatchDrop,
-  CreateDrop,
-  AddDrop,
-  DeleteDrop,
-  RemoveDrop,
-  DeleteAllDrops,
-  RemoveAllDrops,
-  DeleteDropsUnderBucket,
-  RemoveDropsUnderBucket
-} from '@store/actions';
 import { InkDropsService } from '@lib/services';
+import { Action, State, StateContext } from '@ngxs/store';
+import { AddDrop, CreateDrop, DeleteDrop, DeleteDropsUnderBucket, FetchAllDrops, PatchDrop, PopulateAllDrops, RemoveDrop, RemoveDropsUnderBucket, UpdateDrop } from '@store/actions';
+import { map } from 'rxjs/operators';
 
 @State<InkDrops>({
   name: 'drops',
@@ -25,9 +13,9 @@ export class DropsState {
 
   @Action(FetchAllDrops)
   fetchAllDrops(ctx: StateContext<InkDrops>, action: FetchAllDrops) {
-    return this._service.getAll().then(drops => {
+    return this._service.getAll().pipe(map(drops => {
       ctx.dispatch(new PopulateAllDrops(drops));
-    });
+    }));
   }
 
   @Action(PopulateAllDrops)
@@ -37,23 +25,23 @@ export class DropsState {
 
   @Action(UpdateDrop)
   ppdateDrop(ctx: StateContext<InkDrops>, action: UpdateDrop) {
-    return this._service.update(action.dropData).then(drop => {
+    return this._service.update(action.dropData).pipe(map(drop => {
       ctx.dispatch(new PatchDrop((drop as any)._data));
-    });
+    }));
   }
 
   @Action(PatchDrop)
   patchDrop(ctx: StateContext<InkDrops>, action: PatchDrop) {
     const state: any = ctx.getState();
-    const newState = state.map(c => (c._id === action.dropData._id ? action.dropData : c));
+    const newState = state.map(c => (c.id === action.dropData.id ? action.dropData : c));
     ctx.setState([...newState]);
   }
 
   @Action(CreateDrop)
   createDrop(ctx: StateContext<InkDrops>, action: CreateDrop) {
-    return this._service.create(action.dropData).then(drop => {
+    return this._service.create(action.dropData).pipe(map(drop => {
       ctx.dispatch(new AddDrop((drop as any)._data));
-    });
+    }));
   }
 
   @Action(AddDrop)
@@ -64,35 +52,35 @@ export class DropsState {
 
   @Action(DeleteDrop)
   deleteDrop(ctx: StateContext<InkDrops>, action: DeleteDrop) {
-    this._service.delete(action.dropId).then(_ => {
+    this._service.delete(action.dropId).pipe(map(_ => {
       ctx.dispatch(new RemoveDrop(action.dropId));
-    });
+    }));
   }
 
   @Action(RemoveDrop)
   removeDrop(ctx: StateContext<InkDrops>, action: RemoveDrop) {
     const state: any = ctx.getState();
-    const newState = state.filter(b => b._id !== action.dropId);
+    const newState = state.filter(b => b.id !== action.dropId);
     ctx.setState(newState);
   }
 
-  @Action(DeleteAllDrops)
-  deleteAllDrops(ctx: StateContext<InkDrops>, action: DeleteAllDrops) {
-    return this._service.deleteAll().then(_ => {
-      ctx.dispatch(new RemoveAllDrops());
-    });
-  }
+  // @Action(DeleteAllDrops)
+  // deleteAllDrops(ctx: StateContext<InkDrops>, action: DeleteAllDrops) {
+  //   return this._service.deleteAll().then(_ => {
+  //     ctx.dispatch(new RemoveAllDrops());
+  //   });
+  // }
 
-  @Action(RemoveAllDrops)
-  removeAllDrops(ctx: StateContext<InkDrops>, action: RemoveAllDrops) {
-    ctx.setState([]);
-  }
+  // @Action(RemoveAllDrops)
+  // removeAllDrops(ctx: StateContext<InkDrops>, action: RemoveAllDrops) {
+  //   ctx.setState([]);
+  // }
 
   @Action(DeleteDropsUnderBucket)
   deleteDropsUnderBucket(ctx: StateContext<InkDrops>, action: DeleteDropsUnderBucket) {
-    return this._service.deleteAllUnderBucket(action.bucketId).then(bucket => {
+    return this._service.deleteAllUnderBucket(action.bucketId).pipe(map(bucket => {
       ctx.dispatch(new RemoveDropsUnderBucket(action.bucketId));
-    });
+    }));
   }
 
   @Action(RemoveDropsUnderBucket)
