@@ -1,11 +1,12 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Store, Select } from '@ngxs/store';
-import { filter, map } from 'rxjs/operators';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { InkBoardMeta } from '@lib/models';
-import { InkBoardsService, InkDatabaseService } from '@lib/services';
-import { FetchAllBoards, FetchAllBuckets, FetchAllDrops, RemoveAllBuckets, RemoveAllDrops } from '@store/actions';
+import { Select, Store } from '@ngxs/store';
+import { AuthService } from '@services/auth.service';
+import { UserService } from '@services/user.service';
+import { FetchAllBoards, FetchAllBuckets, FetchAllDrops, RemoveAllBuckets, RemoveAllDrops, SaveUser } from '@store/actions';
 import { BoardsState } from '@store/states';
 import { Observable } from 'rxjs';
+import { AngularFireAuth } from '@angular/fire/auth';
 
 @Component({
   selector: 'inkapp-home-page',
@@ -13,9 +14,27 @@ import { Observable } from 'rxjs';
 })
 export class HomePage implements OnInit, OnDestroy {
   @Select(BoardsState) boards$: Observable<InkBoardMeta[]>;
-  constructor(private _store: Store) {}
+  constructor(private _store: Store, private _authService: AuthService, private _userService: UserService) {
+    this._authService.auth().onAuthStateChanged((user) => {
+      if (user) {
+        this._userService.updateUserData(user);
+        this._store.dispatch(new SaveUser({
+          id: user.uid,
+          isAnonymous: user.isAnonymous,
+          email: user.email,
+          displayName: user.displayName,
+          photoURL: user.photoURL
+        }));
+      } else {
+        console.log('err', user);
+      }
+    });
+  }
 
   ngOnInit() {
+    // this._authService.loginAnonymously().subscribe(user => {
+    //   console.log('sub', user);
+    // });
     this._store.dispatch(new FetchAllBoards());
     this._store.dispatch(new FetchAllBuckets());
     this._store.dispatch(new FetchAllDrops());
