@@ -1,5 +1,6 @@
 export { ExportPage } from './export/export.page';
 export { HomePage } from './home/home.page';
+export { NotFoundPage } from './notfound/nofound.page';
 export { RedirectPage } from './redirect/redirect.page';
 export { SettingsPage } from './settings/settings.page';
-export { NotFoundPage } from './notfound/nofound.page';
+
