@@ -1,5 +1,1 @@
-const checkAge = (age) => age + 1;
-
-
-
-let declare CheckAge = function(age: string): string {};
+//
