@@ -1,1 +1,1 @@
-export const INKAPP_VERSION='1.0.0'
+export const INKAPP_VERSION = '1.0.0-beta.1';
